// src/app/Model/CurrencyMaster/currency-master.model.ts
export class CurrencyMasterModel {
    cur_no: any;
    cur_name: any;
    cur_desc: any;
    createdt: any;
    companycode: any;
}

export class ExchageRateModel {
    cur_no: any;
    from_date: any;
    to_date: any;
    ex_rate: any;
    consumable_ex_rate: any;
}

export class CurrencyMasterModalSearch {
    cur_name: any;
    cur_desc: any;
}